VOOK-003: Improving duration field

+++ models/video.go
@@ -13,7 +13,7 @@ type Video struct {
 	Title       string    `json:"title"`
 	Description string    `json:"description"`
 	Link        string    `json:"link"`
-	Duration    int       `json:"duration"`
+	Duration    int64     `json:"duration"`
 	CreatedAt   time.Time `json:"created_at"`
 	UpdatedAt   time.Time `json:"updated_at"`
 }
